risk_score changed to int and stopwatch being stooped

+++ PerimeterXModule/DataContracts/Activities/ActivityDetails.cs
@@ -15,12 +15,12 @@ namespace PerimeterX
 		public string ModuleVersion;
 
 		[DataMember(Name = "block_score")]
-		public double RiskScore;
+		public int RiskScore;
 
 		[DataMember(Name = "pass_reason", EmitDefaultValue = true)]
 		public PassReasonEnum? PassReason;
 
 		[DataMember(Name = "risk_rtt")]
-		public double RiskRoundtripTime;
+		public long RiskRoundtripTime;
 	}
 }
